+++ main.go
@@ -3,7 +3,7 @@ package main
 import "os"
 
 const (
-	VERSION = "0.6.0"
+	VERSION = "0.7.0"
 )
 
 func main() {