+++ stmt/stmt.go
@@ -23,6 +23,7 @@ var (
 	ErrNonAssignable      = errors.New("cannot assign value to non-subscriptable, line %d")
 	ErrExpectedName       = errors.New("expected filename at import, line %d")
 	ErrCannotImport       = errors.New("cannot import outside of global scope, line %d")
+	ErrExpectedInteger    = errors.New("expected expression to be integer, line %d")
 	ErrProgramExit        = errors.New("")
 
 	ErrReturnOutsideFunc = ConditionalError{Msg: "cannot use return outside of a function, line %d", Type: RETURN}
@@ -59,6 +60,7 @@ type Statement struct {
 	Type       int
 	Line       int
 	Operator   int
+	Range 	   int
 	Name       string
 	Params     []string
 	Statements []Statement